rintf -> rint
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@11434 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/af/xap/unix/xap_UnixPSGraphics.cpp
@@ -181,7 +181,7 @@ UT_uint32 PS_Graphics::measureUnRemappedChar(const UT_UCSChar c)
 	// FIXME but we're not smart enough to do that yet
 	float fWidth = m_pCurrentFont->measureUnRemappedChar(c, m_pCurrentFont->getSize())
 		* ((double)getResolution() / (double)getDeviceResolution());
-	return static_cast<UT_uint32>(rintf(fWidth));
+	return static_cast<UT_uint32>(rint(fWidth));
 }
 #endif //#ifndef WITH_PANGO
